Refresh tree after adding or removing project (#7)
* refresh tree after adding or removing project

* Always check if client is defined

* fix lint errors

* remove extra nodejs versions from pipeline

+++ src/commands/database/deleteDocument.ts
@@ -4,13 +4,16 @@ import { DocumentTreeItem } from "../../tree/database/DocumentTreeItem";
 import { confirmDialog } from "../../ui/confirmDialog";
 
 export async function deleteDocument(documentTreeItem: DocumentTreeItem): Promise<void> {
+    if (!databaseClient) {
+        return;
+    }
     const document = documentTreeItem.document;
     const collection = documentTreeItem.parent.parent.collection;
     try {
         const shouldDelete = await confirmDialog(`Delete document "${document["$id"]}" from ${collection.name}?`);
         if (shouldDelete) {
             await databaseClient.deleteDocument(collection.$id, document["$id"]);
-            documentTreeItem.parent.window.showInformationMessage(`Deleted document "${document["$id"]}" from ${collection.name}.`);
+            window.showInformationMessage(`Deleted document "${document["$id"]}" from ${collection.name}.`);
         }
     } catch (e) {
         window.showErrorMessage(e);
